clean up

+++ libs/uk_bpf_runtime/uk_builtin_bpf_helpers.c
@@ -145,7 +145,7 @@ uint64_t bpf_get_addr(const char *function_name)
 
 uint64_t bpf_probe_read(void* addr_to_read, uint64_t size)
 {
-	uint64_t addr = addr_to_read;
+	uint64_t addr = (uint64_t)addr_to_read;
 	if (size != 1 && size != 4 && size != 8) {
 		debug("bpf_probe_read: invalid size %lu\n", size);
 		return 0;
